Initial commit of BantamOS
BantamOS aims to be a real-time preemptive multitasking operating
system for use in Roost Technology's embedded projects.
BantamOS is not optimized for execution speed or memory usage, but is
optimized for code readability and hackability. It is the intention
of the BantamOS developers that anybody with at least some embedded
systems development or Arduino experience should be able to understand
the BantamOS code base with minimal difficulty.

Signed-off-by: ZachWick <zach@zachwick.xxx>